python/requirements.txt
# 高擎电机CAN控制依赖包
python-can>=4.5.0
numpy>=1.21.0